+++ src/index.client.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { hydrateRoot } from 'react-dom/client';
 import { createBrowserRouter } from 'react-router';
 import { RouterProvider } from 'react-router/dom';
-import routes from './routes';
+import routes from '~/routes';
 
 const router = createBrowserRouter(routes, {
   hydrationData: window.__staticRouterHydrationData,
